+++ R/proc_means.R
@@ -262,6 +262,9 @@
 #' The "vardef=" option specifies the variance divisor. Note that the selected
 #' variance divisor can effect the calculations for several statistics.
 #' @param titles A vector of one or more titles to use for the report output.
+#' @param where An expression to filter the rows before the statistics are
+#' calculated. Use the \code{\link[base]{expression}} function to define
+#' the filter.
 #' @return Normally, the requested summary statistics are shown interactively
 #' in the viewer, and output results are returned as a data frame.
 #' If the request produces multiple data frames, they will be returned in a list.
@@ -353,7 +356,8 @@ proc_means <- function(data,
                        class = NULL,
                        weight = NULL,
                        options = NULL,
-                       titles = NULL
+                       titles = NULL,
+                       where = NULL
 ) {
 
   # SAS seems to always ignore these
@@ -515,6 +519,10 @@ proc_means <- function(data,
 
   res <- NULL
 
+  # Deal with where expression
+  if (!is.null(where)) {
+    data <- subset(data, eval(where))
+  }
 
   # Get report if requested
   if (view == TRUE | rptflg) {