+++ R/proc_sort.R
@@ -70,6 +70,8 @@
 #' on the local parameter will override the global setting. The default
 #' value is NULL, which defers to the global setting.  If the global setting
 #' is NULL, the parameter will default to "last".
+#' @param where An expression to filter the rows before the sort takes place.
+#' Use the \code{\link[base]{expression}} function to define the where clause.
 #' @return The sorted dataset.  If a data frame was input, a
 #' data frame will be output.  If a tibble was input, a tibble will
 #' be output.
@@ -136,7 +138,8 @@
 #' @import tibble
 #' @export
 proc_sort <- function(data,  by = NULL, keep = NULL, order = "ascending",
-                      options = NULL, as.character = FALSE, na.sort = NULL) {
+                      options = NULL, as.character = FALSE, na.sort = NULL,
+                      where = NULL) {
 
 
   # Deal with single value unquoted parameter values
@@ -178,6 +181,11 @@ proc_sort <- function(data,  by = NULL, keep = NULL, order = "ascending",
 
   if (nrow(data) == 0) {
     return(data)
+  }
+
+  # Deal with where expression
+  if (!is.null(where)) {
+    data <- subset(data, eval(where))
   }
 
   if (!is.null(order)) {